Feat Filter Marker on the map - 카테고리(지역별)에 해당하는 마커만 랜더링 됩니다. by @frankkim92
