+++ components/detail/detailRight/DetailMap.tsx
@@ -47,7 +47,7 @@ const DetailMap = ({
         style={{
           // 지도의 크기
           width: '100%',
-          height: '200px',
+          height: '230px',
         }}
         level={11} // 지도의 확대 레벨
       >
