+++ firebase.tsx
@@ -4,12 +4,12 @@ import { getFirestore } from 'firebase/firestore';
 import { getStorage } from 'firebase/storage';
 
 const firebaseConfig = {
-  apiKey: 'AIzaSyAm1Oo1wlRi_pfmI4hiS_Z-xSURpgsSrX0',
-  authDomain: 'testpicspot.firebaseapp.com',
-  projectId: 'testpicspot',
-  storageBucket: 'testpicspot.appspot.com',
-  messagingSenderId: '110955419725',
-  appId: '1:110955419725:web:7fad0ffe45eb75af49e404',
+  apiKey: 'AIzaSyCxKx4AXNnQpn6q6fFK6VxvTMEgKChM10o',
+  authDomain: 'picspot-2e239.firebaseapp.com',
+  projectId: 'picspot-2e239',
+  storageBucket: 'picspot-2e239.appspot.com',
+  messagingSenderId: '250271284271',
+  appId: '1:250271284271:web:967a7916cc654f6f39dada',
 };
 
 const app = initializeApp(firebaseConfig);
@@ -65,9 +65,10 @@ export { app, authService, dbService, storageService };
 //   messagingSenderId: "110955419725",
 //   appId: "1:110955419725:web:7fad0ffe45eb75af49e404"
 
-// apiKey: 'AIzaSyAm1Oo1wlRi_pfmI4hiS_Z-xSURpgsSrX0',
-// authDomain: 'testpicspot.firebaseapp.com',
-// projectId: 'testpicspot',
-// storageBucket: 'testpicspot.appspot.com',
-// messagingSenderId: '110955419725',
-// appId: '1:110955419725:web:7fad0ffe45eb75af49e404',
+// picspot33
+// apiKey: "AIzaSyC_hPhusKHFzK9txrOrF3Ry_UDojdG3ehI",
+// authDomain: "picspot33-d70df.firebaseapp.com",
+// projectId: "picspot33-d70df",
+// storageBucket: "picspot33-d70df.appspot.com",
+// messagingSenderId: "1071145662979",
+// appId: "1:1071145662979:web:b0ad278354defe1ffdec36"
